# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:7055
paths:
  /api/v1/documents:
    get:
      operationId: search
      parameters:
      - in: query
        name: pageable
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDocumentDTO"
          description: OK
      tags:
      - document-controller
    post:
      operationId: save
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInsertOrUpdateDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentDTO"
          description: Created
      tags:
      - document-controller
  /api/v1/documents/{id}:
    delete:
      operationId: delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
      tags:
      - document-controller
    get:
      operationId: findById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentDTO"
          description: OK
      tags:
      - document-controller
    put:
      operationId: update
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInsertOrUpdateDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentDTO"
          description: OK
      tags:
      - document-controller
components:
  schemas:
    DocumentDTO:
      type: object
      properties:
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        id:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
    DocumentInsertOrUpdateDTO:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
        title:
          type: string
          maxLength: 250
          minLength: 0
      required:
      - content
      - title
    PageDocumentDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/DocumentDTO"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean